+++ src/data_gradients/managers/abstract_manager.py
@@ -13,7 +13,7 @@ from data_gradients.utils.utils import print_in_box
 from data_gradients.visualize.seaborn_renderer import SeabornRenderer
 from data_gradients.utils.pdf_writer import ResultsContainer, Section, FeatureSummary
 from data_gradients.utils.summary_writer import SummaryWriter
-from data_gradients.datasets.adapter.base_adapter import BaseDatasetAdapter
+from data_gradients.batch_processors.base import BaseDatasetAdapter
 
 logging.basicConfig(level=logging.INFO)
